% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time_ppmify_helpers.R
\name{aggregate_points_space_time}
\alias{aggregate_points_space_time}
\title{The aggregate_points_space_time function}
\usage{
aggregate_points_space_time(
  points,
  ppmx,
  periods,
  date_start_end,
  reference_raster
)
}
\description{
Helper function for space_time_ppmify
}
