% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_batch.R
\name{choose_batch}
\alias{choose_batch}
\title{Adaptively select locations to sample based on uncertainty and covariance. 
The algorithm selects locations with high uncertainy and low correlation}
\usage{
choose_batch(XY, entropy, candidate, rho, nu, batch_size)
}
\arguments{
\item{XY}{candidate locations where the batch is choosen from}

\item{entropy}{entropy computed at the locations XY}

\item{rho}{Matern covariance parameters of the model}

\item{nu}{Matern covariance parameters of the model}

\item{batch_size:}{number of observations in the batch}
}
\value{
Returns the row indeces of the selected coordinates.
}
\description{
Adaptively select locations to sample based on uncertainty and covariance. 
The algorithm selects locations with high uncertainy and low correlation
}
