% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_batch_simple.R
\name{choose_batch_simple}
\alias{choose_batch_simple}
\title{Function to take a spatially disaggregated sample weighted by values at points (e.g. uncertainty)}
\usage{
choose_batch_simple(point_data, batch_size, uncertainty_fieldname, candidate)
}
\arguments{
\item{point_data}{an sf object containing points}

\item{batch_size}{number of sampled points required}

\item{uncertainty_fieldname}{column name of the field by which sampling should be weighted}

\item{candidate}{Boolean vector of length equal to the number of points in `point_data` representing whether a point should be considered a candidate for sampling or not}
}
\value{
index of the row numbers of `point_data` included in the sample
}
\description{
Function to take a spatially disaggregated sample weighted by values at points (e.g. uncertainty)
}
