% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sf_raster.R
\name{dist_to_sf_raster}
\alias{dist_to_sf_raster}
\title{Function create raster of distance to an sf/sp object.}
\usage{
dist_to_sf_raster(sf_obj, ref_raster)
}
\arguments{
\item{sf_obj}{An sf or sp object to calculate distance to}

\item{ref_raster}{a reference raster with the extent and resolution of}
}
\description{
Note that if a line/polygon sf object is provided, it is first rasterized to the 
reference raster and distance to the raster centroid of the cell in which the 
line passes through is used as the line coordinate.
}
