% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hal_parallel.R
\name{fit_hal_parallel}
\alias{fit_hal_parallel}
\title{Helper function to fit a highly adaptive lasso to a fold as part of parallelized call}
\arguments{
\item{folds_df_list_fold}{}

\item{X_var}{Names of column corresponding covariates to use}

\item{n_pos_var}{Name of column corresponding to numbers positive}

\item{n_neg_var}{Name of column corresponding to numbers negative}
}
\description{
Helper function to fit a highly adaptive lasso to a fold as part of parallelized call
}
