% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds_list_to_df_list.R
\name{folds_list_to_df_list}
\alias{folds_list_to_df_list}
\title{Helper function to convert a list of folds to list of dfs}
\usage{
folds_list_to_df_list(fold, df)
}
\arguments{
\item{fold}{}

\item{df}{}
}
\description{
Helper function to convert a list of folds to list of dfs
}
