% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_posterior_metrics.R
\name{gam_posterior_metrics}
\alias{gam_posterior_metrics}
\title{Function to return metrics from posterior of (binomial) GAM model}
\usage{
gam_posterior_metrics(gam_mod, new_data, n_sims, exceedance_threshold)
}
\arguments{
\item{gam_mod}{gam model object}

\item{new_data}{data.frame of new data}

\item{n_sims}{Number of simulations}

\item{exceedance_threshold}{Prevalence threshold to use for exceedance probabilities}
}
\description{
Function to return metrics from posterior of (binomial) GAM model
}
