% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time_ppmify_helpers.R
\name{get_int_points_exposure_weights}
\alias{get_int_points_exposure_weights}
\title{The get_int_points_exposure_weights function}
\usage{
get_int_points_exposure_weights(
  ppmx,
  ppm_cases_points_counts,
  exposure_raster,
  num_periods
)
}
\arguments{
\item{ppmx}{}
}
\description{
Helper function for space_time_ppmify
}
