% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_range.R
\name{optimal_range}
\alias{optimal_range}
\title{Finds optimal range parameter for GP GAM}
\arguments{
\item{y}{vector of outcome}

\item{x}{vector of column names of covariates}

\item{coords_cols}{vector of length 2 referring to column names of x and y coordinates}

\item{min_dist}{min range}

\item{max_dist}{max range}

\item{length.out}{number of possible ranges to try}

\item{model_data}{model data}

\item{m1}{numeric argument (1-5) referring to covariance function to use}

\item{k}{k argument for gam}
}
\description{
Finds optimal range parameter for GP GAM
}
