% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_predictor_mgcv_st.R
\name{prevalence_predictor_mgcv_st}
\alias{prevalence_predictor_mgcv_st}
\title{Function to predict prevalence excedance probabilities at spatial locations. Currently only support binomial data.}
\usage{
prevalence_predictor_mgcv_st(point_data, time_field = NULL,
  layer_names = NULL, v = 10, exceedance_threshold,
  batch_size = NULL, uncertainty_fieldname = NULL,
  additional_covariates = NULL,
  covariate_extractor_url = "https://faas.srv.disarm.io/function/fn-covariate-extractor",
  seed = 1981, m = 3, fix_cov = NULL)
}
\arguments{
\item{point_data}{Required. An sf object of points containing at least `n_trials`, `n_positive` fields}

\item{time_field}{Optional name of column referring to time. Time should be coded as an integer not date.}

\item{layer_names}{Optional names of column corresponding covariates to use. Choose from 'Layer names' as 
outlined [here](https://github.com/disarm-platform/fn-covariate-extractor/blob/master/SPECS.md). If none provided
then spatial only model assumed.}

\item{v}{The number of folds to use in the machine learning step. Defaults to 10.}

\item{exceedance_threshold}{Required. The prevalence threshold for which exceedance probabilities are required}

\item{batch_size}{The number of adaptively selected locations required. Defaults to NULL.}

\item{uncertainty_fieldname}{If `batch_size` is specified (>0), adaptive sampling is performed. To sample
in order to minimize classification uncertainty choose 'exceedance_probability'. To sample in order to minimize prediction
error choose 'prevalence_bci_width'.}

\item{additional_covariates}{Optional vector of column names of `point_data` referencing additional covariates to 
include in the model. Defulats to NULL}

\item{covariate_extractor_url}{The function currently makes use of the temporary DiSARM API function `fn-covariate-extractor` 
to extract values of `layer_names` at locations specified in `point_data`. If this algorithm is hosted 
somewhere other than the DiSARM API, include the URL here.}

\item{m}{check out m in ?gam}
}
\description{
Function to predict prevalence excedance probabilities at spatial locations. Currently only support binomial data.
}
