% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time_ppmify.R
\name{space_time_ppmify}
\alias{space_time_ppmify}
\title{The space_time_ppmify function}
\usage{
space_time_ppmify(points, layer_name = NULL, covariates = NULL,
  exposure, resolution = 1, date_start_end, num_periods = 1,
  prediction_exposure = exposure, approx_num_int_points = 10000,
  prediction_frame = FALSE)
}
\arguments{
\item{points}{sfc object with a `Date` associated with the point in yyyy-mm-dd format}

\item{layer_name}{Optional. Names of the bioclimatic/environmental layers to use as covariates. 
Currently only uses static covariates. 
See [here under 'Layer names'](https://github.com/disarm-platform/fn-covariate-extractor/blob/master/SPECS.md) 
for a list of options. If none provided, spatial only model is assumed.}

\item{covariates}{a rasterLayer or rasterStack of additional covariates to include}

\item{exposure}{rasterLayer of exposure (to be used as offset). Required. 
Raster representing the population over which points arose. Currently 
only accepts a single raster which is used across all time periods.}

\item{resolution}{Resolution in km2 that covariates and offset should be resampled to (>= 1km2). Defaults to 1.}

\item{date_start_end}{Required. Vector of 2 values representing the start and end 
times over which points were observed in yyyy-mm-dd format}

\item{num_periods}{Number of time periods over which to aggregate points, 
where 1 considers all points in a single time period 
(equivalent to assuming a spatial only model). Defaults to 1.}

\item{prediction_exposure}{Optional rasterLayer of exposure to be used for prediction. 
This may be different to `exposure` if `points` arose from a different population 
to that you wish to predict to. Automatically resampled to the same 
resolution as `exposure`. Defaults to `exposure`.}

\item{prediction_frame}{Logical. Whether the function should also return a 
data frame required for prediction (i.e. raster cells)? Defaults to FALSE.}

\item{density}{Required. Density of quadrature points to generate (points / square km)}
}
\value{
A data.frame object with the following fields:
\itemize{
  \item x - x coordinates
  \item y - y coordinates
  \item exposure - Population for that point in space and time to be used as offset (when logged)
  \item period - Number 1 through number of layers as determined by `date_start_end` and `num_periods`
  \item outcome - Whether the point is an observation (1) or a quadrature point (0)
  \item regression_weights - Number of `outcomes` per space-time cell, to be used as a regression weight
}
}
\description{
This function builds off the ppmify function from Nick Golding's ppmify package, 
converting your case data into a data.frame suitable for applying Poisson regression models.
}
