% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sample_auto.R
\name{adaptive_sample_auto}
\alias{adaptive_sample_auto}
\title{Identical to adaptive.sample from geosample package, except auto decreases delta if supplied value too large}
\usage{
adaptive_sample_auto(obj1, obj2, pred.var.col = NULL,
  excd.prob.col = NULL, batch.size = 1, delta, criterion,
  poly = NULL, plotit = TRUE)
}
\arguments{
\item{obj1}{candidates}

\item{obj2}{existing sample}

\item{pred.var.col}{pred.var.col}

\item{excd.prob.col}{excd.prob.col}

\item{batch.size}{batch size}

\item{delta}{min distance}

\item{criterion}{'exceedprob' or 'predvar'}

\item{poly}{poly}

\item{plotit}{logical}
}
\description{
Identical to adaptive.sample from geosample package, except auto decreases delta if supplied value too large
}
