% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sf_raster.R
\name{dist_to_sf_raster}
\alias{dist_to_sf_raster}
\title{Function create raster of distance to an sf/sp object. Note that the
function currently only calcualtes distances to coordinates/vertices of the 
sf/sp object rather than distance to nearest section/line. 
If this is a line or polygon object this could result in 
some introduced error depending on the density of vertices}
\usage{
dist_to_sf_raster(sf_obj, ref_raster)
}
\arguments{
\item{sf_obj}{An sf or sp object to calculate distance to}

\item{ref_raster}{a reference raster with the extent and resolution of}
}
\description{
Function create raster of distance to an sf/sp object. Note that the
function currently only calcualtes distances to coordinates/vertices of the 
sf/sp object rather than distance to nearest section/line. 
If this is a line or polygon object this could result in 
some introduced error depending on the density of vertices
}
