% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_predictor_mgcv.R
\name{prevalence_predictor_mgcv}
\alias{prevalence_predictor_mgcv}
\title{Function to predict prevalence excedance probabilities at spatial locations. Currently only support binomial data.}
\usage{
prevalence_predictor_mgcv(point_data, layer_names, v = 10,
  exceedance_threshold, batch_size, uncertainty_fieldname)
}
\arguments{
\item{point_data}{An sf object of points containing `n_trials`, `n_positive` fields}

\item{layer_names}{Names of column corresponding covariates to use. Choose from 'Layer names' as outlined [here](https://github.com/disarm-platform/fn-covariate-extractor/blob/master/SPECS.md)}

\item{v}{The number of folds to use in the machine learning step. Defaults to 10.}

\item{exceedance_threshold}{The prevalence threshold for which exceedance probabilities are required}

\item{batch_size}{The number of adaptively selected locations required}
}
\description{
Function to predict prevalence excedance probabilities at spatial locations. Currently only support binomial data.
}
