% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_map.R
\name{quick_map}
\alias{quick_map}
\title{Function to make quick leaflet maps of spatial data}
\usage{
quick_map(
  plot_data,
  value_field = NULL,
  colors = NULL,
  circle_size = 3,
  raster_legend_title = NULL
)
}
\arguments{
\item{plot_data}{An sp, sf or rasterLayer object to plot}

\item{value_field}{Names of column corresponding to values to plot}

\item{colors}{Optional set of hexcolors to create a color palette from}
}
\description{
Function to make quick leaflet maps of spatial data
}
