% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_plus_close_pairs.R
\name{spatial_sample_plus_close_pairs}
\alias{spatial_sample_plus_close_pairs}
\title{Take a spatially dispersed sample from a set of 2D points, with optional inclusion of close-pairs of points. 
Designed for estimation of spatial covariance.}
\usage{
spatial_sample_plus_close_pairs(coords, n_close, n_spatial)
}
\arguments{
\item{coords}{A matrix or data frame of 2 columns representing x and y coordinates of points to sample from.}

\item{n_close}{Optional. The number of closely separated points to select. Must be multiple of 5.}

\item{n_spatial}{Optional. The number of spatially dispersed points to select.}
}
\value{
Returns a list with 'sample' - the coordinates of selected points and 'sample_idx' - the row indeces of the selected coordinates.
}
\description{
Take a spatially dispersed sample from a set of 2D points, with optional inclusion of close-pairs of points. 
Designed for estimation of spatial covariance.
}
\examples{
coords <- data.frame(x = runif(1000,0,1), y = runif(1000,0,1))
samp <- spatial_sample_plus_close_pairs(coords, n_close = 50, n_spatial = 50)
plot(coords, col="gray"); points(coords[samp,], pch=16)
}
\keyword{hexbin}
